backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
